requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml --resolver=backtracking --all-extras -o requirements.txt
affine==2.4.0
    # via rasterio
alabaster==1.0.0
    # via sphinx
attrs==25.3.0
    # via rasterio
azure-core==1.35.0
    # via azure-storage-blob
azure-storage-blob==12.25.1
    # via ocha-stratus
babel==2.17.0
    # via sphinx
beautifulsoup4==4.13.4
    # via furo
certifi==2025.7.9
    # via
    #   netcdf4
    #   pyogrio
    #   pyproj
    #   rasterio
    #   requests
cffi==1.17.1
    # via cryptography
cfgv==3.4.0
    # via pre-commit
cftime==1.6.4.post1
    # via netcdf4
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via
    #   click-plugins
    #   cligj
    #   dask
    #   rasterio
click-plugins==1.1.1.2
    # via rasterio
cligj==0.7.2
    # via rasterio
cloudpickle==3.1.1
    # via dask
cryptography==45.0.5
    # via azure-storage-blob
dask==2025.5.1
    # via ocha-stratus
distlib==0.3.9
    # via virtualenv
docutils==0.21.2
    # via
    #   myst-parser
    #   sphinx
filelock==3.18.0
    # via virtualenv
fsspec==2025.5.1
    # via dask
furo==2024.8.6
    # via ocha-lens (pyproject.toml)
geopandas==1.1.1
    # via ocha-stratus
identify==2.6.12
    # via pre-commit
idna==3.10
    # via requests
imagesize==1.4.1
    # via sphinx
iniconfig==2.1.0
    # via pytest
isodate==0.7.2
    # via azure-storage-blob
jinja2==3.1.6
    # via
    #   myst-parser
    #   ocha-stratus
    #   sphinx
locket==1.0.0
    # via partd
markdown-it-py==3.0.0
    # via
    #   mdit-py-plugins
    #   myst-parser
markupsafe==3.0.2
    # via jinja2
mdit-py-plugins==0.4.2
    # via myst-parser
mdurl==0.1.2
    # via markdown-it-py
myst-parser==4.0.1
    # via ocha-lens (pyproject.toml)
netcdf4==1.7.2
    # via ocha-lens (pyproject.toml)
nodeenv==1.9.1
    # via pre-commit
numpy==2.3.1
    # via
    #   cftime
    #   geopandas
    #   netcdf4
    #   pandas
    #   pyogrio
    #   rasterio
    #   rioxarray
    #   shapely
    #   xarray
ocha-stratus==0.1.3
    # via ocha-lens (pyproject.toml)
packaging==25.0
    # via
    #   dask
    #   geopandas
    #   pyogrio
    #   pytest
    #   rioxarray
    #   sphinx
    #   xarray
pandas==2.3.1
    # via
    #   ocha-lens (pyproject.toml)
    #   geopandas
    #   ocha-stratus
    #   xarray
partd==1.4.2
    # via dask
platformdirs==4.3.8
    # via virtualenv
pluggy==1.6.0
    # via pytest
pre-commit==4.2.0
    # via ocha-lens (pyproject.toml)
psycopg2-binary==2.9.10
    # via ocha-stratus
pyarrow==20.0.0
    # via ocha-stratus
pycparser==2.22
    # via cffi
pygments==2.19.2
    # via
    #   furo
    #   pytest
    #   sphinx
pyogrio==0.11.0
    # via geopandas
pyparsing==3.2.3
    # via rasterio
pyproj==3.7.1
    # via
    #   geopandas
    #   rioxarray
pytest==8.4.1
    # via ocha-lens (pyproject.toml)
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.1
    # via ocha-lens (pyproject.toml)
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via
    #   dask
    #   myst-parser
    #   pre-commit
rasterio==1.4.3
    # via rioxarray
requests==2.32.4
    # via
    #   azure-core
    #   sphinx
rioxarray==0.19.0
    # via ocha-stratus
roman-numerals-py==3.1.0
    # via sphinx
ruff==0.12.3
    # via ocha-lens (pyproject.toml)
shapely==2.1.1
    # via geopandas
six==1.17.0
    # via
    #   azure-core
    #   python-dateutil
snowballstemmer==3.0.1
    # via sphinx
soupsieve==2.7
    # via beautifulsoup4
sphinx==8.2.3
    # via
    #   ocha-lens (pyproject.toml)
    #   furo
    #   myst-parser
    #   sphinx-basic-ng
    #   sphinx-copybutton
sphinx-basic-ng==1.0.0b2
    # via furo
sphinx-copybutton==0.5.2
    # via ocha-lens (pyproject.toml)
sphinxcontrib-applehelp==2.0.0
    # via sphinx
sphinxcontrib-devhelp==2.0.0
    # via sphinx
sphinxcontrib-htmlhelp==2.1.0
    # via sphinx
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-qthelp==2.0.0
    # via sphinx
sphinxcontrib-serializinghtml==2.0.0
    # via sphinx
sqlalchemy==2.0.41
    # via ocha-stratus
toolz==1.0.0
    # via
    #   dask
    #   partd
typing-extensions==4.14.1
    # via
    #   azure-core
    #   azure-storage-blob
    #   beautifulsoup4
    #   sqlalchemy
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via requests
virtualenv==20.31.2
    # via pre-commit
xarray==2025.7.1
    # via
    #   ocha-lens (pyproject.toml)
    #   ocha-stratus
    #   rioxarray
